:white_check_mark:/:hammer: Refactor data files path
